forget decimal and make it double!

+++ MarketDomain/Prominence.cs
@@ -10,8 +10,8 @@ namespace MarketDomain
             return $"Ticker: {Ticker}\tPrice {Price}\tChange Amount {ChangeAmount}";
         }
         public string Ticker { get; set; } = string.Empty;
-        public decimal? Price { get; set; }
-        public decimal? ChangeAmount { get; set; }
+        public double? Price { get; set; }
+        public double? ChangeAmount { get; set; }
         public string? ChangePercentage { get; set; }
         public long? Volume { get; set; }
         public int Id { get; set; }
