+++ TurtleAPI/AlphaVantage/AlphaVProminenceResult.cs
@@ -6,8 +6,8 @@ namespace TurtleAPI.AlphaVantage
     public class AlphaVProminenceResult
     {
         public string ticker { get; set; } = string.Empty;
-        public decimal? price { get; set; }
-        public decimal? change_amount { get; set; }
+        public double? price { get; set; }
+        public double? change_amount { get; set; }
         public string? change_percentage { get; set; }
         public long? volume { get; set; }
     }
